Moving some files around and fixing imports.  Installs now!
